Artificial-Intelligence/AIWritingIndicator-13/requirements/tools.txt
GitPython==3.1.46
gitdb==4.0.12
smmap==5.0.2
filelock==3.20.3
fsspec==2026.1.0
python-docx==1.2.0